Hello World!!!
package main: This line declares that the package name is main. In Go, every executable program must have a main package.
import “fmt”: This line imports the fmt package, which contains functions for formatting text, including printing to the console.
func main(): This line declares the main function. The main function is the entry point of a Go program. When the program runs, the code inside this function is executed.
fmt.Println(“Hello, World!”): This line calls the Println function from the fmt package to print the string “Hello, World!” followed by a newline to the console.