Project-2
This code makes the computer choose a random number between 1 and 100 and keeps guessing until the user guesses correctly. After each guess, the user is told whether their guess should be larger or smaller. If he/she makes a correct guess and wants to play the game again, he/she should press the 1 button. If he doesn't want to play, press the 2 button.